move generate_sp_tokenizer to functional.py.
