+++ docs/source/data/utils.rst
@@ -11,8 +11,3 @@ torchtext.data.utils
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. autofunction:: ngrams_iterator 
-
-:hidden:`generate_sp_tokenizer`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: generate_sp_tokenizer
